# feat: implement ONNX optimization and lazy loading for SentenceTransformer (#34)