--- a/pyproject.toml
+++ b/pyproject.toml
@@ -36,6 +36,9 @@ dependencies = [
     "mcp[cli]>=0.3.0",
     "gitignore-parser>=0.1.11", # Support for respecting .gitignore files
     "xdg-base-dirs>=6.0.2",
+    "onnx>=1.18.0",
+    "onnxruntime>=1.22.0",
+    "optimum>=1.25.3",
 ]
 
 [project.scripts]